# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_beuphecan_remote-time-tracker_internal_dto.CheckAdminExistsResponse:
    properties:
      exists:
        type: boolean
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.InitAdminRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.InitAdminResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      user:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse'
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@remotetimetracker.com
    name: API Support
  description: API documentation for Remote Time Tracker - A comprehensive time tracking
    application with screenshot capture, task management, organization/workspace management,
    and synchronization features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Remote Time Tracker API
  version: "1.0"
paths:
...
  /system/admin-exists:
    get:
      description: Returns whether a system admin has been created. Used to determine
        if initial setup is required.
      produces:
      - application/json
      responses:
        "200":
          description: Admin status checked
          schema:
            allOf:
            - $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.CheckAdminExistsResponse'
              type: object
        "500":
          description: Failed to check admin status
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      summary: Check if system admin exists
      tags:
      - system
  /system/init-admin:
    post:
      consumes:
      - application/json
      description: Creates the first system admin. This endpoint only works when no
        admin exists in the system. Used for initial setup.
      parameters:
      - description: Admin details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.InitAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: System admin created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.InitAdminResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "409":
          description: Admin already exists
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "500":
          description: Failed to create admin
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      summary: Initialize system admin
      tags:
      - system
  /system/uploads/check:
    get:
      description: Verifies that upload and screenshot folders exist and are writable
      produces:
      - application/json
      responses:
        "200":
          description: Upload folder check completed
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check uploads folder health
      tags:
      - system
  /system/uploads/ensure:
    post:
      description: Creates necessary upload folders (uploads, screenshots) if they
        don't exist
      produces:
      - application/json
      responses:
        "200":
          description: Upload folders ensured
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "500":
          description: Failed to create folders
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ensure upload folders exist
      tags:
      - system
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...